rename adapter to parser
