+++ src/webapp/services/file-parser/FileParser.ts
@@ -5,9 +5,9 @@ import _ from "lodash";
 export const FILE_MAPPER = "files.json";
 export const FILES_FOLDER = "files";
 
-export interface FileAdapter<T> {
+export interface FileParser<T> {
     prefix: string;
-    parse(files: FileEntry[]): FutureData<T[]>;
+    fromEntity(files: FileEntry[]): FutureData<T[]>;
     toEntries(data: T[]): FutureData<FileEntry[]>;
 }
