+++ src/webapp/services/file-parser/models/FileEntry.ts
