+++ src/webapp/services/file-parser/models/FileMap.ts
