+++ src/webapp/services/file-parser/models/TranslatableTextModel.ts
